Disgusting quick code for testing SSBO vs 3D texture for raycast speed.
Unfortunately, my contact at Nvidia was correct in predicting that SSBOs have a substantial speed penalty over 3D textures, due to the loss of optimal tiling (among other reasons).

Somehow, I'll have to find a solution that combines the lookup speed of textures with the upload performance of persistent-mapped SSBOs.